refactor(Preconditions): move package `global.error` -> `global.util`
